rename EventScheduling to ActionScheduling
